Fix randomNumber not avoiding current scenario
randomNumber() was not taking into account lasRanNum when generating a
new number allowing the app to re-roll the same scenario. Current fix
avoids that while maintaining an even probability for all scenarios.
In order to implement this, it was more convenient to switch to a zero-
based index affecting code in a few other places.

Other changes:
- The dropdown is modified to make "Choose a Scenario" unselectable.
- The dropdown is automatically reset when rolling a new scenario to
avoid the visual inconsistency.